[chore][pkg/stanza] Cleanup csv parser operator files
